+++ lib/cocoapods/target/aggregate_target.rb
@@ -288,10 +288,10 @@ module Pod
 
     # @param  [String] config_name The build configuration name to get the xcconfig for
     # @return [String] The path of the xcconfig file relative to the root of
-    #         the Pods project.
+    #         the user project.
     #
-    def xcconfig_relative_path(config_name)
-      relative_to_pods_root(xcconfig_path(config_name)).to_s
+    def xcconfig_relative_srcroot(config_name)
+      xcconfig_path(config_name).relative_path_from(client_root).to_s
     end
 
     # @return [String] The path of the copy resources script relative to the